Move amqplib_compat modules to top level
Fixes #38
